Fixed merge issues related to balance var in Player class

+++ app/src/main/java/com/example/rrcasino/Player.java
@@ -63,7 +63,7 @@ public class Player {
         return balance;
     }
 
-    public void addToBalance(double cash) {
+    public void addToBalance(float cash) {
         balance += cash;
     }
